crashlogd: fix event type enum and wd_array definitions
print_eventtype is realigned with event type enum.

wd_array must be of same size of event type enum in the current design
as wd_array is walked using event type enum. This is due to bad design
as it should not have correlation. FULL_REPORT is removed from wd_array
as a workaround.

Due to this workaround, all folders are always watched as full wd_array
is initialised. Still, callback function are not set thanks to
FULL_REPORT if-def in main when setting callbacks, thus no action will
happen in case of change in those folders.

Change-Id: I6a5f191de4532789ade7579f4b4eaa05866a5b04
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>
Reviewed-on: https://android.intel.com/275385
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Fouassier, Yann <yann.fouassier@intel.com>