amtl: Add Cherrytrail SSIC configuration
This patch add a specific AMTL configuration to
handle the BP logging in case IPC SSIC is selected.

As well the patch modify the path of storing the
bplogs which have been updated into a encrypted
partition.
bplogs are now stored under /data/logs/ partition.
New location is applicable only to IMIN/GMIN.

Change-Id: If374f991f633e5da9afae7e57a80d0a61c6052d6
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-9943
Signed-off-by: Petru Muraru <petrux.muraru@intel.com>
Signed-off-by: Nicolas LOUIS <nicolasx.louis@intel.com>
Reviewed-on: https://android.intel.com:443/303010
Reviewed-on: https://android.intel.com:443/303857