[BUGFIX] log_service: do not start ap[k]_logfs services if /data is not decrypted
Starting ap_logfs or apk_logfs before /data is decrypted implies that
cryptfs will need to stop those services before decrypting. This
action is triggered by setting "persist.service.aplogfs.enable"
property to 1, set by default.prop in ramdisk, in early boot stage.

If encrypted, /data is mounted twice, 1st as tmpfs, 2nd as decrypted.
Property "vold.post_fs_data_done" only ensure that /data is mounted
and post-fs-data actions are done, not that it is decrypted.

In the case that /data is encrypted, check that framework was started,
this means "late_start" class of services, the kind used for
ap_log_srv, is started. This ensure that the ap_logfs and apk_logfs
will be started only when /data is decrypted.

Change-Id: I627a93beffbd7848f1c7a09a34f44ee20372b3f9
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-26255
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>
Reviewed-on: https://android.intel.com:443/344740