crashlog: update folder reset procedure on clean startup
permissions on paths are not correctly set from crashlog on clean
startup because of mkdir and chown.
This patch aims to clean only the contents of given paths rather
then fully removing and recreating them.
The path permissions are to be updated mainly through init files.
Heap profiling and core events are filtered based on value set for
property persist.core.enabled.

Change-Id: I6610049334b45f7f288d05080b0e35e8a046d7d2
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-34424
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: https://android.intel.com:443/361618